Adding support for permissions based access control
Rationale: Having to modify code (by writing methods for every new role) is
not ideal in cases where it is required to create roles on the fly (e.g.
through an admin panel). For these cases, it is helpful to have permission
based methods so that response depends on whatever permissions a user has.